requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.52
botocore==1.29.52
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.0
Deprecated==1.2.13
Django==4.1.5
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-oauth-toolkit==2.2.0
django-restframework==0.0.1
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.14.0
flower==1.2.0
humanize==4.4.0
idna==3.4
jmespath==1.0.1
jwcrypto==1.4.2
kombu==5.2.4
oauthlib==3.2.2
pendulum==2.1.2
Pillow==9.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
pycparser==2.21
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
pytzdata==2020.1
redis==4.4.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
wrapt==1.14.1
